add pbh sidecar
