+++ world-chain-builder/crates/world/rpc/src/lib.rs
@@ -3,8 +3,8 @@
 pub mod sequencer;
 pub use sequencer::SequencerClient;
 
-pub mod transactions;
-pub use transactions::EthTransactionsExt;
+pub mod api_ext;
+pub use api_ext::EthTransactionsExt;
 
 pub mod core;
 pub use core::{EthApiExtServer, WorldChainEthApiExt};
